pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = ""
authors = [
    {name = "4vpr",email = "4shout4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.119.1,<0.120.0)",
    "tortoise-orm[psycopg] (>=0.25.1,<0.26.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
